backend/requirements.txt
anyio==3.3.1
asgiref==3.2.10
astroid==2.4.2
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
charset-normalizer==2.0.5
coverage==5.3
dadata==20.7.0
decorator==4.4.2
dj-environ==0.0.1
Django==3.1.1
django-cors-headers==3.5.0
django-filter==2.3.0
djangorestframework==3.11.1
gunicorn==20.0.4
h11==0.9.0
httpcore==0.10.2
httpx==0.14.3
idna==2.10
ipython==7.18.1
ipython-genutils==0.2.0
isort==5.5.2
jedi==0.17.2
lazy-object-proxy==1.5.1
lxml==4.7.1
Markdown==3.2.2
mccabe==0.6.1
parso==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
prompt-toolkit==3.0.7
psycopg2-binary==2.8.6
ptyprocess==0.6.0
Pygments==2.7.1
pylint==2.6.0
python-docx==0.8.11
pytz==2020.1
requests==2.24.0
rfc3986==1.4.0
six==1.15.0
sniffio==1.1.0
sqlparse==0.3.1
toml==0.10.1
traitlets==5.0.4
urllib3==1.25.10
wcwidth==0.2.5
wrapt==1.12.1
